requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
Deprecated==1.2.13
Django==3.2.13
django-environ==0.8.1
django-extensions==3.1.5
django-redis==5.2.0
djangorestframework==3.13.0
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.7
pytz==2022.1
redis==4.2.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
zipp==3.7.0
